+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [3.0.0-beta.150](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.149...v3.0.0-beta.150) (2023-06-27)
+
+
+### Bug Fixes
+
+* **plugin-meetings:** set breakoutStatus cache false avoid duplicate changes ([#2900](https://github.com/webex/webex-js-sdk/issues/2900)) ([a5c9e5e](https://github.com/webex/webex-js-sdk/commit/a5c9e5e2d852620d9c4d42fa25cd6b0dfb28fcb1))
+
+## [3.0.0-beta.149](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.148...v3.0.0-beta.149) (2023-06-26)
+
+
+### Features
+
+* **plugin-meetings:** added setActiveSpeakerCsis() API ([#2899](https://github.com/webex/webex-js-sdk/issues/2899)) ([80dfafd](https://github.com/webex/webex-js-sdk/commit/80dfafd8fd64d07d3cfd3d9ae77d59e533860550))
+
+## [3.0.0-beta.148](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.147...v3.0.0-beta.148) (2023-06-26)
+
+
+### Bug Fixes
+
+* **Live-Annotation:** fix annotation info is null ([#2901](https://github.com/webex/webex-js-sdk/issues/2901)) ([447f10b](https://github.com/webex/webex-js-sdk/commit/447f10bebc42d14e2599467c98f7f71d053bf3d4))
+
 ## [3.0.0-beta.147](https://github.com/webex/webex-js-sdk/compare/v3.0.0-beta.146...v3.0.0-beta.147) (2023-06-23)